refactor: made utility global
